requirements.txt
Flask
Flask-Restful
Flask-SQLAlchemy
Flask-JWT-Extended